requirements.txt
Flask==3.1.2
mysql_connector_repackaged==0.3.1
nltk==3.9.2
numpy==2.4.1
scikit_learn==1.8.0
sentence_transformers==5.2.0
